- unify type handling by introducing new Type class that transparently defers conversion of the syntax-level type to the concrete llvm assembly type until code generation and frees the ast nodes from having to deal with syntax-level types - fix cleanup in update-grammar script - create module builder in treeparser and directly work with that instead of returning pointers to function and struct vectors and instantiating the module builder after the fact

+++ update-grammar.sh
@@ -88,10 +88,14 @@
 		mv $GRAMMAR_OUT/*.cc "$GRAMMAR_DST_C"
 		mv $GRAMMAR_OUT/*.h "$GRAMMAR_DST_H"
 		
-		# cleanup
-		rm -rf "$GRAMMAR_OUT"
-		
 		echo "OK"
-		exit 0
  	fi
  fi
+ 
+ if test -d "$GRAMMAR_OUT" ; then
+ 	rm -rf "$GRAMMAR_OUT"
+ fi
+ 
+ exit 0
+ 
+
